move makeRecentRepositoriesGroup to utility file
